Make generation of an image tarball a common task
Adds `CreateImageTarball`, based on GCE's `CreateTarball` but naming the
tarball file after the parameters defined on the manifest.
